Rename test namespace so that we can use psalm-internal classes within tests
